+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased - ReleaseDate
 
+* Add `Anchr template` scripts
+    * 0_cleanup.sh
+    * 0_real_clean.sh
+
 ## 0.1.5 - 2020-10-24
 
 * Use `tsv-sample` to replace `shuf`