+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 ## Unreleased - ReleaseDate
 
 * Add `Anchr template` scripts
-    * 0_cleanup.sh
-    * 0_real_clean.sh
-    * 0_master.sh
+  * 0_cleanup.sh
+  * 0_real_clean.sh
+  * 0_master.sh
+  * 4_down_sampling.sh
 
 ## 0.1.5 - 2020-10-24