+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased - ReleaseDate
 
+## 0.2.0 - 2020-10-26
+
 * Binary name `anchr`
 
 * Add `anchr template` scripts